👌 move RUM-slim entry point
