+++ .eslintrc.js
@@ -252,7 +252,7 @@ module.exports = {
       },
     },
     {
-      files: ['packages/{rum,logs,rum-slim}/src/index.ts', 'packages/rum/src/entries/*.ts'],
+      files: ['packages/{rum,logs,rum-slim}/src/entries/*.ts'],
       rules: {
         'local-rules/disallow-enum-exports': 'error',
       },
