+++ packages/rum-slim/package.json
@@ -2,9 +2,9 @@
   "name": "@datadog/browser-rum-slim",
   "version": "3.8.0",
   "license": "Apache-2.0",
-  "main": "cjs/index.js",
-  "module": "esm/index.js",
-  "types": "cjs/index.d.ts",
+  "main": "cjs/entries/main.js",
+  "module": "esm/entries/main.js",
+  "types": "cjs/entries/main.d.ts",
   "scripts": {
     "build": "run-p build:cjs build:esm build:bundle",
     "build:bundle": "rm -rf bundle && webpack --mode=production",
